server: Prefer RTC route from RR client
In case that a Route Reflector(RR) and a non RR client peering, peering
of different RR clusters for example, the RR should send the RTC route
from its client even if the RTC route from non RR client is better path
based on the best path algorithm in order to notify that some RR clients
are interested in the given Route Target.

Currently, only source peer address is concerned, the RTC route from RR
client can have lower priority than non RR client and it can not be
advertised.

This patch fixes to prefer the route from RR client when selecting the
candidate to be advertised.

Signed-off-by: IWASE Yusuke <iwase.yusuke0@gmail.com>